ej6/requirements.txt
langchain
langgraph
google-generativeai
Pillow
jupyter
ipython
matplotlib
python-dotenv
pydantic>=2.0.0
